// src/app/models/surveyresults.ts
/* tslint:disable */
export interface SURVEYRESULTS {
  ALLIEVO_REPARTO?: string;
  ID?: number;
  USER_ID?: number;
  CREATED?: string;
  MODIFIED?: string;
  CORSO?: string;
  DATA?: string;
  DOCENTE?: string;
  ALLIEVO_NOME?: string;
  ALLIEVO_COGNOME?: string;
  ALLIEVO_EMAIL?: string;
  ALLIEVO_AZIENDA?: string;
  UNIQ_ID?: string;
  Q1?: string;
  Q2?: string;
  Q3?: string;
  Q4?: string;
  Q5?: string;
  Q6?: string;
  Q7?: string;
  Q8?: string;
  Q9?: string;
  Q10?: string;
  Q11?: string;
  Q12?: string;
}